modify path and config
