--- react-vite/src/data/NavLinkItem.ts
@@ -1,10 +1,0 @@
-export const NavLinkItem = [
-  {
-    url: "home",
-    label: "Home",
-  },
-  {
-    url: "contact",
-    label: "Contact",
-  },
-];
