+++ react-vite/src/pages/main.tsx
@@ -4,23 +4,13 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import "./main.css";
 import ErrorPage from "./error/ErrorPage";
 import Root from "../routes/Root";
-import Home from "./home/Home";
-import Contact from "./contact/Contact";
+import { browserRouterChildrenData } from "../data/NavLinkItem";
 const router = createBrowserRouter([
   {
     path: "/",
     element: <Root />,
     errorElement: <ErrorPage />,
-    children: [
-      {
-        path: "/home",
-        element: <Home />,
-      },
-      {
-        path: "/contact",
-        element: <Contact />,
-      },
-    ],
+    children: browserRouterChildrenData,
   },
 ]);
