+++ react-vite/tsconfig.json
@@ -5,6 +5,10 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
+    "baseUrl": "src",
+    "paths": {"@/*": [
+      "*"
+    ]},
 
     /* Bundler mode */
     "moduleResolution": "bundler",
